+++ voucher-impl/src/main/java/com/dwarfeng/voucher/impl/bean/entity/HibernateCheckerInfo.java
@@ -17,12 +17,16 @@ public class HibernateCheckerInfo implements Bean {
 
     private static final long serialVersionUID = 8363527331527179417L;
 
-    // -----------------------------------------------------------主键-----------------------------------------------------------
+    // region 主键
+
     @Id
     @Column(name = "id", length = Constraints.LENGTH_STRING_ID, nullable = false, unique = true)
     private String stringId;
 
-    // -----------------------------------------------------------主属性字段-----------------------------------------------------------
+    // endregion
+
+    // region 主属性字段
+
     @Column(name = "enabled", nullable = false)
     private boolean enabled;
 
@@ -35,14 +39,20 @@ public class HibernateCheckerInfo implements Bean {
     @Column(name = "remark", length = Constraints.LENGTH_REMARK)
     private String remark;
 
-    // -----------------------------------------------------------多对一-----------------------------------------------------------
+    // endregion
+
+    // region 多对一
+
     @OneToOne(targetEntity = HibernateVoucherCategory.class)
     @JoinColumns({ //
             @JoinColumn(name = "id", referencedColumnName = "id", insertable = false, updatable = false), //
     })
     private HibernateVoucherCategory voucherCategory;
 
-    // -----------------------------------------------------------审计-----------------------------------------------------------
+    // endregion
+
+    // region 审计
+
     @DatamarkField(handlerName = "checkerInfoDatamarkHandler")
     @Column(
             name = "created_datamark",
@@ -58,10 +68,13 @@ public class HibernateCheckerInfo implements Bean {
     )
     private String modifiedDatamark;
 
+    // endregion
+
     public HibernateCheckerInfo() {
     }
 
-    // -----------------------------------------------------------映射用属性区-----------------------------------------------------------
+    // region 映射用属性区
+
     public HibernateStringIdKey getKey() {
         return Optional.ofNullable(stringId).map(HibernateStringIdKey::new).orElse(null);
     }
@@ -70,7 +83,10 @@ public class HibernateCheckerInfo implements Bean {
         this.stringId = Optional.ofNullable(idKey).map(HibernateStringIdKey::getStringId).orElse(null);
     }
 
-    // -----------------------------------------------------------常规属性区-----------------------------------------------------------
+    // endregion
+
+    // region 常规属性区
+
     public String getStringId() {
         return stringId;
     }
@@ -134,6 +150,8 @@ public class HibernateCheckerInfo implements Bean {
     public void setModifiedDatamark(String modifiedDatamark) {
         this.modifiedDatamark = modifiedDatamark;
     }
+
+    // endregion
 
     @Override
     public String toString() {